Tar Arch Linux package
This is because the artifact API has some restrictions on valid
characters for filenames. The Spotify package in particular makes use
of epoch which is then represented as `:epoch` in the filename. And
`:` just happens to be one of the invalid characters.

Even though it's a pain to unpack twice (artifacts can't be downloaded
individually and are just packed into a single `.zip`) later on, it
seems like the only way so far.